Split invariant checks to hourly and daily
